# packages/core/api/openapi.yaml
openapi: 3.0.3
info:
  title: The Traitors UK API
  description: A RESTful API providing structured data for The Traitors UK series 1-4, sourced from Wikipedia.
  version: 1.0.0
servers:
  - url: /api
    description: Firebase Cloud Functions Proxy
paths:
  /series:
    get:
      summary: List all series
      description: Returns a list of all available series of The Traitors UK.
      operationId: listSeries
      responses:
        "200":
          description: A successful response with an array of series.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Series"
  /series/{seriesId}:
    get:
      summary: Get a specific series
      description: Returns detailed information for a single series.
      operationId: getSeriesById
      parameters:
        - name: seriesId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: A successful response with the series data.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Series"
        "404":
          description: Series not found.
  /series/{seriesId}/candidates:
    get:
      summary: Get candidates for a series
      description: Returns a list of all candidates who participated in a specific series.
      operationId: getCandidatesBySeries
      parameters:
        - name: seriesId
          in: path
          required: true
          schema:
            type: integer
        - name: limit
          in: query
          description: The maximum number of candidates to return.
          schema:
            type: integer
            minimum: 1
            default: 25
        - name: offset
          in: query
          description: The number of candidates to skip before starting to collect the result set.
          schema:
            type: integer
            minimum: 0
            default: 0
        - name: sortBy
          in: query
          description: The field to sort by.
          schema:
            type: string
            enum: [name]
            default: name
        - name: sortOrder
          in: query
          description: The order to sort by.
          schema:
            type: string
            enum: [asc, desc]
            default: asc
      responses:
        "200":
          description: A successful response with an array of candidates.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Candidate"
        "404":
          description: Series not found.
  /series/{seriesId}/votes:
    get:
      summary: Get votes for a series
      description: Returns a list of all votes cast in a specific series.
      operationId: getVotesBySeries
      parameters:
        - name: seriesId
          in: path
          required: true
          schema:
            type: integer
        - name: limit
          in: query
          description: The maximum number of votes to return.
          schema:
            type: integer
            minimum: 1
            default: 20
        - name: offset
          in: query
          description: The number of votes to skip before starting to collect the result set.
          schema:
            type: integer
            minimum: 0
            default: 0
      responses:
        "200":
          description: A successful response with an array of votes.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Vote"
        "404":
          description: Series not found.
  /candidates/search:
    get:
      summary: Search for candidates by name
      description: Returns a list of candidates whose names match the search query.
      operationId: searchCandidates
      parameters:
        - name: name
          in: query
          required: true
          description: The name of the candidate to search for.
          schema:
            type: string
      responses:
        "200":
          description: A successful response with an array of matching candidates.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Candidate"
        "400":
          description: Invalid request, name parameter is missing.
  /candidates/{candidateId}:
    get:
      summary: Get a specific candidate
      description: Returns detailed information for a single candidate.
      operationId: getCandidateById
      parameters:
        - name: candidateId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: A successful response with the candidate data.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Candidate"
        "404":
          description: Candidate not found.

components:
  schemas:
    Series:
      type: object
      properties:
        id:
          type: integer
          description: The unique identifier for the series.
          example: 1
        year:
          type: integer
          description: The year the series was broadcast.
          example: 2022
        title:
          type: string
          description: The full title of the series.
          example: "The Traitors (UK series 1)"
      required:
        - id
        - year
        - title
    Candidate:
      type: object
      properties:
        id:
          type: integer
          description: The unique identifier for the candidate.
          example: 101
        name:
          type: string
          description: The name of the candidate.
          example: "Alyssa"
        seriesId:
          type: integer
          description: The ID of the series the candidate participated in.
          example: 1
        role:
          type: string
          description: The candidate's role in the game.
          enum: [Faithful, Traitor]
        outcome:
          type: string
          description: The final outcome for the candidate.
          example: "Murdered"
      required:
        - id
        - name
        - seriesId
        - role
        - outcome
    Vote:
      type: object
      properties:
        id:
          type: integer
          description: The unique identifier for the vote event.
        episode:
          type: integer
          description: The episode number in which the vote occurred.
        voterId:
          type: integer
          description: The ID of the candidate who cast the vote.
        votedForId:
          type: integer
          description: The ID of the candidate who received the vote.
        seriesId:
          type: integer
          description: The ID of the series this vote belongs to.
      required:
        - id
        - episode
        - voterId
        - votedForId
        - seriesId